fix: juniorPromisePage 상단 타이틀 배너 컴포넌트로 분리
